+++ timetracker.py
@@ -1,258 +1,188 @@
+import cmd
+import sys
 import json
-import time
 import os.path
+import time
+from time import strftime
 from datetime import date
 
-#TODO
-# - Poikkeavan Inputin hallinta
-# - 
-# - Koulukurssin yksilollinen seuranta (Tyoprojektien seuranta)
-# - Datan syotto ulos
-#   - Komentorivi <- DONE
-#   - Harvest
+class timeObject():
+  start = None
+  end = None
+  date = None
+  timeUsed = None
+  client = None
+  project = None
 
-time_types = ["sleep", "study", "work", "other"]
-time_types_lyh =  ["un", "op", "ty", "mu"]
-time_end_types = ["loppu", "end", "l", "lounas"]
-close_types = "exit"
-help_type = "help"
-nayta_list = ["n", "nayta", "show"]
+  def __init__(self, start, client='ceili', project='muu'):
+    self.start = start
+    self.client = client
+    self.project = project
+    self.date = date.today()
 
-default_project = "muu"
-default_client = 'ceili'
-default_projects = {
-    "project_name" : default_project,
-    "time_used" : 0
-}
+  def countUsedTime(self):
+    """ Counts used time to minutes
+    """
+    self.timeUsed = int(round((self.end-self.start)/60,6))
 
-clients = {
-    "client" : default_client,
-    "client_projects" : [default_projects]
-}
+  def toJSON(self):
+    """ Returns object in Dict format
+    """
+    return {"start": str(self.start) ,"end": str(self.end),"date": str(self.date),"timeUsed": str(self.timeUsed),"client": self.client,"project": self.project}
 
+  def timeToStr(self, val):
+    """ Parses seconds to str
+    """
+    return strftime('%H:%M:%S', time.localtime(val))
 
-default_day = { 
-    "date" : str(date.today()),
-    "sleep" : 0,
-    "study" : 0,
-    "work" : [clients],
-    "other" : 0
-}
-
-data_folder = "data/2021"
-print(os.getcwd())
-if os.getcwd() == "/Users/otto":
-        data_folder = "coding/timetracker/data/2021"
-
-def timetracker():
-    data = initWeek()
-    readTimeType(data)
+  def printTimeObject(self):
+    """ Prints objcet to cmd/terminal
+    """
+    print()
+    #print('Date: ' + str(self.date))
+    #printTime('Start time:', self.start)
+    printTime('End time:', self.end)
+    print('Client: ' + self.client)
+    print('Project: ' + self.project)
+    print('Used time: ' + minToHoursAndMins(self.timeUsed))
+    print()
 
 
-def initWeek():
-    weekNum = str(getWeek())
-    if os.path.isfile(data_folder+'week'+weekNum+'.json'):
-        with open(data_folder+'week'+weekNum+'.json') as file_data:
-            print('File already exists')
-            data = json.load(file_data)
-            inc_today = True
-            for day in data:
-                if day["date"] == str(date.today()):
-                    inc_today = False
-            if inc_today:
-                data.append(default_day)
+class TimetrackerShell(cmd.Cmd):
+  intro = 'Welcome to timetracker. Type help or ? to list commands.\n'
+  prompt = '$ '
 
-            return data
-    else:
-        with open(data_folder+'week'+weekNum+'.json', 'w') as file_data:
-            print('Creating new file / starting new week')
-            data = []
-            data.append(default_day)
-            file_data.write(json.dumps(data, indent=4))
-            file_data.close()
-            return data
-            
+  filePath = 'testdata/' # data file path
+  if "timetracker" not in os.getcwd():
+    filePath = "coding/timetracker/testdata/"
+  data = None # list of time tracking objects
+  curTracking = None # time tracking object to handle attributes
 
-def readTimeType(data):
-    time_type = input().split(' ')
-
-    client = ''
-    project = ''
-
-    if (len(time_type) > 2):
-        client = time_type[1]
-        project = time_type[2]
-    elif (len(time_type) == 2 ):
-        client = 'ceili'
-        project = time_type[1]
-
-    if close_types == time_type[0]:
-        close()
-    if help_type == time_type[0]:
-        help_info()
-    if time_type[0] in nayta_list:
-        nayta(data)
-    for i in range(0,len(time_types)):
-        if time_types[i] == time_type[0]:
-            if (len(time_type) > 1):
-                startTime(data, time_types[i], client, project)
-            else: startTime(data, time_types[i], default_client, default_project)
-        elif time_types_lyh[i] == time_type[0]:
-            if (len(time_type) > 1):
-                startTime(data, time_types[i], client, project)
-            else: startTime(data, time_types[i], default_client, default_project)
+  # --- timetracker commands ---
+  def do_ty(self, arg):
+    'Start time tracking for work'
+    params = parse(arg)
+    self.startTime(params)
     
 
-def startTime(data, type_t, client, project):
-    timeStart = time.time()
+  def do_l(self, arg):
+    'End time tracking'
+    if self.curTracking is None:
+      return
+    self.endTime()
+
+  def do_show(self, arg):
+    'Show current weeks time usage'
+    printData(self.data)
+
+  def do_exit(self, arg):
+    'Stop application'
+    if self.curTracking is not None:
+      self.endTime()
+    print('Ending time tracking')
+    return True
+
+  # --- utility methods ---
+
+  def startTime(self, params):
+    """ Actions to start time tracking
+    """
+    if self.curTracking is not None:
+      self.handleEndTracking()
+      self.curTracking = None
+    self.curTracking = timeObject(time.time(), params[0], params[1])
+    self.prompt = 'Client: '+ self.curTracking.client + '\nProject: ' + self.curTracking.project + '\n$ '
     print()
-    print("Started " + type_t)
-    print("Client: " + client)
-    print("Project: " + project)
-    timePrint("", timeStart)
-    endTime(data, timeStart, type_t, client, project)
+    printTime('Start time:' , self.curTracking.start)
 
+  def endTime(self):
+    """ Actions after ending time tracking
+    """
+    self.handleEndTracking()
+    self.curTracking.printTimeObject()
+    self.prompt = '$ '
+    self.curTracking = None
 
-def endTime(data, timeStart, type_t, client, project):
-    listen = True
-    while listen:
-        input_listen = input()
-        if input_listen == close_types:
-            timeEnd = time.time()
-            usedTime = countUsedTime(timeEnd, timeStart)
-            saveData(data, usedTime, type_t, client, project)
-            listen = False
-        if input_listen in time_end_types:
-            timeEnd = time.time()
-            print()
-            timePrint("Start time: " ,timeStart)
-            timePrint("End time: " ,timeEnd)
-            print()
-            print("Used time: ")
-            usedTime = countUsedTime(timeEnd, timeStart)
-
-            #Debug
-            #usedTime = 420
-
-            listen = False
-            saveData(data, usedTime, type_t, client, project)
-            readTimeType(data)
-        if input_listen in nayta_list:
-            timeEnd = time.time()
-            usedTime = countUsedTime(timeEnd, timeStart)
-
-            #Debug
-            #usedTime = 420
-
-            listen = False
-            saveData(data, usedTime, type_t, client, project)
-            nayta(data)
-
-
-
-def saveData(data, usedTime, type_t, client, project):
-    weekNum = str(getWeek())
-    if os.path.isfile(data_folder+'week'+weekNum+'.json'):
-        with open(data_folder+'week'+weekNum+'.json', 'w') as file_data:
-            for day in data:
-                if day["date"] == str(date.today()):
-                    if type_t == "work":
-                        clientExists = False
-                        projectExists = False
-                        for cli in day[type_t]:
-                            if cli["client"] == client:
-                                for proj in cli["client_projects"]:
-                                    if proj["project_name"] == project:
-                                        proj["time_used"] = proj["time_used"] + usedTime
-                                        projectExists = True
-                                if not projectExists:
-                                    newProject = {
-                                        "project_name" : project,
-                                        "time_used" : usedTime
-                                    }
-                                    cli["client_projects"].append(newProject)
-                                clientExists = True
-                        if not clientExists:
-                            newProject = {
-                                "project_name" : project,
-                                "time_used" : usedTime
-                            }
-                            newClient = {
-                                "client" : client,
-                                "client_projects" : [newProject]
-                            }
-                            day[type_t].append(newClient)
-                    else:
-                        day[type_t] = day[type_t] + usedTime
-            file_data.write(json.dumps(data, indent=4))
-            file_data.close()
-    else:
-        print('File not found')
-
-
-def countUsedTime(timeEnd, timeStart):
-    timeUsed = int(round((timeEnd-timeStart)/60, 6))
-    timeUsedHour = int(timeUsed // 60)
-    timeUsedMin = int(timeUsed%60)
-    print()
-    print(str(timeUsedHour)+" hour and "+str(timeUsedMin)+" min")
-    print()
-    return timeUsed #Palauttaa ajan minuutteina
+  def handleEndTracking(self):
+    self.curTracking.end = time.time()
+    self.curTracking.countUsedTime()
+    self.data.append(self.curTracking)
     
 
-def timePrint(info ,timePrint):
-    timePrint = time.localtime(timePrint)
-    print(info + str(timePrint[3])+":"+str(timePrint[4]))
+  # --- data methods ---
+  def postloop(self):
+    """ Saves the data to correct JSON-file
+    """
+    curWeekFilePath = curWeekPath(self.filePath)
+    with open(curWeekFilePath, 'w') as file_data:
+      data = []
+      for i in self.data:
+        data.append(i.toJSON())
+      file_data.write(json.dumps(data, indent=4))
+      file_data.close()
 
 
-def getWeek():
-    return date.today().isocalendar()[1]
+  def preloop(self):
+    """ Reads the data from JSON-file and parses it to editable objects
+    """
+    curWeekFilePath = curWeekPath(self.filePath)
+    if os.path.isfile(curWeekFilePath):
+      with open(curWeekFilePath, 'r') as file_data:
+        self.data = []
+        data = json.load(file_data)
+        for i in data:
+          startTime = float(i['start'])
+          endTime = float(i['end'])
+          obj = timeObject(startTime, i['client'], i['project'])
+          obj.date = date.fromisoformat(i['date'])
+          obj.end = endTime
+          obj.timeUsed = int(i['timeUsed'])
+          self.data.append(obj)
+    else:
+      self.data = []
 
+
+def printData(data):
+  print()
+  indent = "  "
+  dataDict = {}
+  for i in data:
+    dataDict[i.client] = {}
+  for i in data:
+    dataDict[i.client][i.project] = 0
+  for i in data:
+    dataDict[i.client][i.project] = dataDict[i.client][i.project] + i.timeUsed
+  for i in dataDict.keys():
+    print(i)
+    for j in dataDict[i]:
+      print(indent + j + ' : ' + minToHoursAndMins(dataDict[i][j]))
+  print()
+
+def curWeekPath(folderPath):
+  """ Return filename and path
+  """
+  weekNum = str(date.today().isocalendar()[1])
+  year = str(date.today().year)
+  return folderPath + '/' + year + '_week'+weekNum+'.json'
+
+def parse(arg):
+  'Parses args to list'
+  return arg.split()
+
+def printTime(text ,argTime):
+  """ Prints time (seconds) with text
+  """
+  localTime = time.localtime(argTime)
+  strTime = str(localTime[3])+':'+str(localTime[4])
+  print(text + ' ' + strTime)
+  return strTime
 
 def minToHoursAndMins(mins):
-    timeUsedHour = int(mins // 60)
-    timeUsedMin = int(mins%60)
-    return str(timeUsedHour)+" hour and "+str(timeUsedMin)+" min"
-
-def nayta(data):
-    #print(json.dumps(data, indent=4, sort_keys=False))
-    sis = "    "
-    print("-------------------------------------------------------")
-    for day in data:
-        overal_worktime = 0
-        print("Date: " + day["date"])
-        for cli in day["work"]:
-            print(sis + cli["client"])
-            for proj in cli["client_projects"]:
-                overal_worktime = overal_worktime + proj["time_used"]
-                print(sis + sis + proj["project_name"] + " : " + minToHoursAndMins(proj["time_used"]))
-        print()
-        print("   Overal worktime: "+ minToHoursAndMins(overal_worktime))
-        print("-------------------------------------------------------")
-    readTimeType(data)
-
-
-def help_info():
-    print()
-    print("Start time keeping:")
-    print()
-    for i in range(0,len(time_types)):
-        print(time_types[i] + " - " + time_types_lyh[i])
-    print()
-    print("End time keeping:")
-    print()
-    for i in range(0, len(time_end_types)):
-        print(time_end_types[i])
-    
-    print()
-    print("Exit app: exit")
-    timetracker()
-        
-
-def close():
-    print("Time tracking ended!")
-    exit()
-
+  """ Formats minutes to hours and minutes
+  """
+  timeUsedHour = int(mins // 60)
+  timeUsedMin = int(mins%60)
+  return str(timeUsedHour)+" hour and "+str(timeUsedMin)+" min"
 
 if __name__ == '__main__':
-    timetracker()
+  TimetrackerShell().cmdloop()
